+++ tests/test_model_methods.py
@@ -285,12 +285,15 @@ def test_setup_reservoirs_no_control(
     example_wflow_model.tables.write()
 
     test_table = example_wflow_model.tables.data[f"reservoir_sh_{lake_id}"]
+    test_table_hq = example_wflow_model.tables.data[f"reservoir_hq_{lake_id}"]
 
     example_wflow_model.tables.data.clear()
     example_wflow_model.root.set(new_root, mode="r")
     example_wflow_model.tables.read()
+    tables_data = example_wflow_model.tables.data
 
-    assert example_wflow_model.tables.data[f"reservoir_sh_{lake_id}"].equals(test_table)
+    assert tables_data[f"reservoir_sh_{lake_id}"].equals(test_table)
+    assert tables_data[f"reservoir_hq_{lake_id}"].equals(test_table_hq)
 
 
 @pytest.mark.timeout(120)  # max 2 min